Add --wide and --ignore-errors Options To Use of Managing Metadata in Git
Add `--wide` option to `profile tidy` and `application tidy` to flatten
CustomApplication and Profile metadata so that the high-level elements
such as profileActionOverrides and fieldPermissions are stored on a
single line along with their children.

Grouping related elements into a single line makes it easier to resolve
merge conflicts when the metadata is stored in git.

Add --ignore-errors option to leave tidied metadata unchanged if it
can't be parsed.  This allows tidy to fail gracefully when there are
conflict markers in the metadata.

These two options can be combined to store metadata in a flattened, wide
format in git, but keep it in the normal, long format in your working
copy.

To use with git, configure clean and smudge git filters to use force-md:
$ git config --local filter.salesforce-application.clean 'force-md application tidy --wide --ignore-errors -'
$ git config --local filter.salesforce-application.smudge 'force-md application tidy --ignore-errors -'
$ git config --local filter.salesforce-profile.clean 'force-md profile tidy --wide --ignore-errors -'
$ git config --local filter.salesforce-profile.smudge 'force-md profile tidy --ignore-errors -'

Update .gitattributes to use the filters:
*.app-meta.xml filter=salesforce-application
*.profile-meta.xml filter=salesforce-profile